+++ docs/CHANGELOG.md
@@ -14,6 +14,10 @@ nav_order: 5
 
     *Andrew Schwartz*
 
+* Add Rails engine support to generators.
+
+    *Tomasz Kowalewski*
+
 ## 3.14.0
 
 * Defer to built-in caching for language environment setup, rather than manually using `actions/cache` in CI.